Connect TUIO and JavaFX
Make the gestures accept a node as sort-of listener. It's not super
pretty, but oh well. Had to also search for nodes manually since JavaFX
decided not to support something simple for it. Hope it is enough.

There are now 5 events: swipe up/down/left/right and tap. Tap is
converted to a mouse click event for JavaFX, the rest are swipe events.
Note that the mouse click event has click count 1.